get system information in JS, not in Haskell
The library in Haskell does not work very well on Windows. It looks
like the JS alternative is a lot more solid. So that logic was moved
into the JS code, and it should provide better quality data.

This also means we do not need to ask people how many GB of RAM they
have anymore.