output to build.log / make upload instructions
I think the best way to gather the data is to have people open GitHub
issues on the repo. That means the file extension has to be one of a
couple options, and I needed to do some style tweaks to make sure the
format of the overview looks alright.